complete the ERC20Bridge (is no accurate yet), the project is virtually stable in this commit, prepare to implement the the ETH bridge for GoldenBridge smart contract

+++ test/GoldenBridgeTest.t.sol
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.20;
 
-import { Test } from "forge-std/Test.sol";
-import { TokenL2 } from "../src/TokenL2.sol";
-import { TokenL1 } from "../src/TokenL1.sol";
-import { IERC20 } from "@openzeppelin/contracts/interfaces/IERC20.sol";
-import { DeployTokenOnL2Script } from "../script/DeployTokenOnL2.s.sol";
-import { console } from "forge-std/console.sol";
-import { GoldenBridge } from "../src/GoldenBridge.sol";
+import {Test} from "forge-std/Test.sol";
+import {TokenL2} from "../src/TokenL2.sol";
+import {TokenL1} from "../src/TokenL1.sol";
+import {IERC20} from "@openzeppelin/contracts/interfaces/IERC20.sol";
+import {DeployTokenOnL2Script} from "../script/DeployTokenOnL2.s.sol";
+import {console} from "forge-std/console.sol";
+import {GoldenBridge} from "../src/GoldenBridge.sol";
 
 contract GoldenBridgeTest is Test {
     address public caller = makeAddr("caller");
@@ -28,7 +28,6 @@ contract GoldenBridgeTest is Test {
     }
 
     function testSetup() public {
-
-        assertEq(uint(1), uint(1));
+        assertEq(uint256(1), uint256(1));
     }
 }
